reading encoder tick and use it in PID regulators, working version, extern is in usage
