+++ Modules/motors/motor_controller.c
@@ -9,7 +9,6 @@
 #include "motor_controller.h"
 #include "control_consts.h"
 
-/* Proponujemy zrobic funkcje updatespeed ktora bedzie */
 
 int16_t PIDSpeedController(float referenceSpeed, float actualSpeed,
 		float currentRegOut) {
