+++ Src/main.c
@@ -8,6 +8,7 @@
 
 #include <stm32f4xx_hal.h>
 #include "motors/timers.h"
+#include "motors/motor_interface.h"
 
 void SysTick_Handler(void){
 	HAL_IncTick();
@@ -16,11 +17,15 @@ void SysTick_Handler(void){
 int main(void) {
 	HAL_Init();
 	InitTimers();
+//	struct singleMotorParam param[1];
+//
+//	param[0].id = LM;
+//	param[0].speed = 50;
 
 //	motor_calibration(CHANNEL2);
-
 	/* Loop forever */
 	while (1) {
+//	setOneSideSpeeds(param,1);
 //	PWM_SetDutyCycle(CHANNEL2, 850);
 
 	}
